Remove mfnav file only if exists - we don't create this file when logged in as admin - or when logged in as user who don't have any investment accounts - it was showing oserror exception for no such file or directory - this change will check to see if the file exists before deleting it

+++ helper.py
@@ -52,7 +52,8 @@ def mfNAV2File():
 
 # Remove MF NAV file upon logout
 def removeMFNAVFile():
-  os.remove(app.config['MFNAV_FILE'])
+  if os.path.exists(app.config['MFNAV_FILE']):
+    os.remove(app.config['MFNAV_FILE'])
 
 # Get Nav of the given MF scheme code
 def getNAV(code):
